gpt_oss/requirements-gpt-oss.txt
# GPT-OSS 本地推理服務依賴包
# 針對RTX 4070 12GB優化

# 核心推理框架
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.6.0

# GPU優化
flash-attn>=2.3.0
xformers>=0.0.22
triton>=2.1.0

# 模型服務
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# 模型量化和優化
auto-gptq>=0.4.0
optimum>=1.14.0

# 工具庫
numpy>=1.24.0
torch-audio>=2.1.0
tokenizers>=0.14.0
safetensors>=0.4.0

# 監控和日誌
prometheus-client>=0.18.0
psutil>=5.9.0
GPUtil>=1.4.0

# HTTP和通信
httpx>=0.25.0
websockets>=12.0
aiofiles>=23.2.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 開發和調試工具
rich>=13.6.0
typer>=0.9.0